Remove references to Tcp Dtos from Event Filters
This allows the EventFilter to be moved into the Storage project without
depending on types in Core.

- Move method that parses Tcp Dto filters to Storage IEventFilters into
TcpClientMessageDtoExtensions.
- Move method used to parse the HTTP filters into the AtomController.
