+++ src/EventStore.Core/Services/Transport/Tcp/ClientTcpDispatcher.cs
@@ -233,7 +233,7 @@ namespace EventStore.Core.Services.Transport.Tcp {
 			var dto = package.Data.Deserialize<TcpClientMessageDto.FilteredReadAllEvents>();
 			if (dto == null) return null;
 
-			IEventFilter eventFilter = EventFilter.Get(dto.Filter);
+			IEventFilter eventFilter = dto.Filter.ToEventFilter();
 
 			int maxSearchWindow = dto.MaxCount;
 			if (dto.MaxSearchWindow.HasValue) {
@@ -269,7 +269,7 @@ namespace EventStore.Core.Services.Transport.Tcp {
 			var dto = package.Data.Deserialize<TcpClientMessageDto.FilteredReadAllEvents>();
 			if (dto == null) return null;
 
-			IEventFilter eventFilter = EventFilter.Get(dto.Filter);
+			IEventFilter eventFilter = dto.Filter.ToEventFilter();
 
 			int maxSearchWindow = dto.MaxCount;
 			if (dto.MaxSearchWindow.HasValue) {
@@ -308,7 +308,7 @@ namespace EventStore.Core.Services.Transport.Tcp {
 			var dto = package.Data.Deserialize<TcpClientMessageDto.FilteredSubscribeToStream>();
 			if (dto == null) return null;
 
-			IEventFilter eventFilter = EventFilter.Get(dto.Filter);
+			IEventFilter eventFilter = dto.Filter.ToEventFilter();
 
 			return new ClientMessage.FilteredSubscribeToStream(Guid.NewGuid(), package.CorrelationId, envelope,
 				connection.ConnectionId, dto.EventStreamId, dto.ResolveLinkTos, user, eventFilter,
